Apply current sender authorization when loading alerts
Note: to address the case where alerts would only get loaded if they were valid at the time of initial parsing and thus stored in leveldb as valid, but the sender was later ignored for example.

+++ src/omnicore/omnicore.cpp
@@ -2988,6 +2988,10 @@ void CMPTxList::LoadAlerts(int blockHeight)
             PrintToLog("ERROR: While loading alert %s: levelDB type mismatch, not an alert.\n", txid.GetHex());
             continue;
         }
+        if (!CheckAlertAuthorization(mp_obj.getSender())) {
+            PrintToLog("ERROR: While loading alert %s: sender is not authorized to send alerts.\n", txid.GetHex());
+            continue;
+        }
 
         if (mp_obj.getAlertType() == 65535) { // set alert type to FFFF to clear previously sent alerts
             DeleteAlerts(mp_obj.getSender());
